Replace `println` calls with enforcable signature
Having the `log` call return a `List<String>` instead of `void` has several
advantages. It makes the printers testable. The `buffer` field was not
accessable in the tests. It also decouples printers from the main logger. Yet an
other advantage is that, instead of imploring implementers of custom printer
classes to use `println` we now enforce it via the type interface. This leads to
less confusion.