Rename DebugFilter to DevelopmentFilter
The name `DebugFilter` often caused confusing where users thought it would
actually display all log messages. While Flutter uses the term 'debug' build
this is not used in the wider Dart community. Especially when talking about a
logs the name 'debug' is overloaded with the old class name. The new name
reflects clearly when you should be using this filter.